fix annotations and add git revision to the program version
